+++ idea-plugin/src/main/kotlin/com/flowable/atlas/generate/dto/DataObjectDtoService.kt
@@ -1,5 +1,6 @@
 package com.flowable.atlas.generate.dto
 
+import com.flowable.atlas.AtlasNotifications.GROUP_ID
 import com.flowable.atlas.index.FlowableModelIndexService
 import com.flowable.atlas.intention.DataObjectBeanGenerator
 import com.flowable.atlas.model.ModelType
@@ -220,7 +221,6 @@ class DataObjectDtoService(private val project: Project) {
         ApplicationManager.getApplication().invokeLater { notify(title, message, type) }
 
     companion object {
-        private const val GROUP_ID = "Flowable Atlas"
 
         fun getInstance(project: Project): DataObjectDtoService = project.service()
     }